refactor: extract helpers from AbstractTransformableImp (977→367 lines)
Extract three helper classes to reduce AbstractTransformableImp.java:
- VehicleManager: static utility for StandIn pool + calculateTurnToFaceAxes (105 lines)
- TransformOperations: instance helper for place/distance/position methods (225 lines)
- TransformAnimator: instance helper for all animate* + OrientationData classes (614 lines)

VantagePointData/PreSetVantagePointData and animateVantagePoint remain on ATI
as protected subclass API used by SymmetricPerspectiveCameraImp and VrUserImp.

All 185 tests pass (119 original + 66 new characterization/contract tests).
Fix 6 directional distance contract tests: point entities (StandInImp) return
signed distances, not always non-negative.

Resolves #581
