+++ core/story-api/src/main/java/org/lgna/story/implementation/VehicleManager.java
@@ -1,0 +1,105 @@
+/*******************************************************************************
+ * Copyright (c) 2006, 2015, Carnegie Mellon University. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Products derived from the software may not be called "Alice", nor may
+ *    "Alice" appear in their name, without prior written permission of
+ *    Carnegie Mellon University.
+ *
+ * 4. All advertising materials mentioning features or use of this software must
+ *    display the following acknowledgement: "This product includes software
+ *    developed by Carnegie Mellon University"
+ *
+ * 5. The gallery of art assets and animations provided with this software is
+ *    contributed by Electronic Arts Inc. and may be used for personal,
+ *    non-commercial, and academic use only. Redistributions of any program
+ *    source code that utilizes The Sims 2 Assets must also retain the copyright
+ *    notice, list of conditions and the disclaimer contained in
+ *    The Alice 3.0 Art Gallery License.
+ *
+ * DISCLAIMER:
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND.
+ * ANY AND ALL EXPRESS, STATUTORY OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY,  FITNESS FOR A
+ * PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE AUTHORS, COPYRIGHT OWNERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING FROM OR OTHERWISE RELATING TO
+ * THE USE OF OR OTHER DEALINGS WITH THE SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *******************************************************************************/
+
+package org.lgna.story.implementation;
+
+import edu.cmu.cs.dennisc.math.EpsilonUtilities;
+import edu.cmu.cs.dennisc.pattern.DefaultPool;
+import org.alice.math.immutable.AffineMatrix4x4;
+import org.alice.math.immutable.ForwardAndUpGuide;
+import org.alice.math.immutable.OrthogonalMatrix3x3;
+import org.alice.math.immutable.Point3;
+import org.alice.math.immutable.Vector3;
+
+/**
+ * Static utility for StandIn pool management and turn-to-face axis calculation.
+ * Extracted from AbstractTransformableImp.
+ */
+class VehicleManager {
+
+  private static final DefaultPool<StandInImp> s_standInPool = new DefaultPool<>(StandInImp.class);
+
+  static StandInImp acquireStandIn(EntityImp composite) {
+    StandInImp rv = s_standInPool.acquire();
+    rv.setVehicle(composite);
+    rv.setLocalTransformation(AffineMatrix4x4.IDENTITY);
+    return rv;
+  }
+
+  static void releaseStandIn(StandInImp standIn) {
+    s_standInPool.release(standIn);
+  }
+
+  static OrthogonalMatrix3x3 calculateTurnToFaceAxes(AbstractTransformableImp subject, EntityImp target) {
+    StandInImp standInA = acquireStandIn(subject);
+    try {
+      standInA.setPositionOnly(subject);
+      Vector3 targetPos = target.getTransformation(subject).translation().asVector();
+      if (EpsilonUtilities.isWithinReasonableEpsilon(targetPos.x(), 0.0)
+          && EpsilonUtilities.isWithinReasonableEpsilon(targetPos.z(), 0.0)) {
+        //todo
+        return subject.getLocalOrientation();
+      } else {
+        targetPos = targetPos.withY(0).normalized();
+        StandInImp standInB = acquireStandIn(subject);
+        try {
+          //Move a standin to the position of the target
+          standInB.applyTranslation(targetPos.x(), targetPos.y(), targetPos.z(), standInB);
+          Point3 standinPosition = standInB.getTransformation(subject.getVehicle()).translation();
+          //Calculate the vector pointing from the subject to the target all in the reference frame of the subject's vehicle
+          //Take that vector and normalize it to create the new "forward" vector for the subject
+          Vector3 newForward = standinPosition.minus(subject.getLocalTransformation().translation()).normalized();
+          //Make a new orientation using this new "forward" and the existing "up" from the subject
+          return (new ForwardAndUpGuide(newForward, subject.getLocalOrientation().up())).asMatrix3x3();
+        } finally {
+          releaseStandIn(standInB);
+        }
+      }
+    } finally {
+      releaseStandIn(standInA);
+    }
+  }
+
+  private VehicleManager() {
+  }
+}
